fix: move useAuth hook to its own file
