+++ src/features/auth/context/useAuth.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react"
+import { AuthContext } from "./auth-context"
+
+export const useAuth = () => {
+  const ctx = useContext(AuthContext)
+  if (!ctx) throw new Error("useAuth debe usarse dentro de AuthProvider")
+  return ctx
+}
